requirements.txt
Pillow>=9.0.0
PyQt5>=5.15.0
PyQt5-Qt5>=5.15.0
PyQt5-sip>=12.8.0